Move the telemetry code out of the CLI package
