+++ packages/telemetry/.eslintrc.json
@@ -1,0 +1,5 @@
+{
+  "extends": [
+    "../../eslintrc.js"
+  ]
+}
