+++ packages/telemetry/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@appland/telemetry",
+  "private": true,
+  "version": "0.1.0",
+  "license": "MIT",
+  "scripts": {
+    "test": "jest",
+    "lint": "eslint telemetry.ts"
+  },
+  "dependencies": {
+    "applicationinsights": "^2.1.4",
+    "conf": "^10.0.2"
+  },
+  "devDependencies": {
+    "@types/jest": "^27.4.1",
+    "@typescript-eslint/eslint-plugin": "^5.7.0",
+    "@typescript-eslint/parser": "^5.7.0",
+    "eslint": "^8.4.1",
+    "jest": "^27.4.7",
+    "sinon": "^13.0.1",
+    "ts-jest": "^27.1.4",
+    "typescript": "^4.5.4"
+  }
+}
