+++ tsconfig.json
@@ -2,5 +2,6 @@
   "extends": "@tsconfig/recommended/tsconfig.json",
   "compilerOptions": {
     "types": ["node", "jest"],
+    "resolveJsonModule": true
   },
 }
